pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "azure-operator"
version = "0.1.0"
description = "Kubernetes-style operator for Azure Landing Zones using Bicep/ARM"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "Platform Team", email = "platform@example.com"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: System :: Systems Administration",
]
# SECURITY: Dependencies pinned to tested minor versions.
# Update only after explicit compatibility testing.
# Run `pip-audit` and `trivy` before any version bump.
dependencies = [
    "pydantic>=2.5.0,<2.11.0",
    "azure-identity>=1.15.0,<1.20.0",
    "azure-mgmt-resource>=23.0.0,<23.3.0",
    "azure-core>=1.29.0,<1.33.0",
    "PyYAML>=6.0.1,<6.1.0",
    "click>=8.1.0,<8.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0,<8.4.0",
    "pytest-asyncio>=0.21.0,<0.26.0",
    "pytest-cov>=4.1.0,<6.1.0",
    "mypy>=1.7.0,<1.15.0",
    "types-PyYAML>=6.0.12,<6.1.0",
    "ruff>=0.1.6,<0.10.0",
    "python-dotenv>=1.0.0,<1.1.0",
    "pip-audit>=2.6.0,<2.8.0",
    "safety>=2.3.0,<3.3.0",
]

[project.scripts]
azure-operator = "controller.main:run"
azo = "controller.cli:main"

[project.urls]
Homepage = "https://github.com/example/azure-operator"
Repository = "https://github.com/example/azure-operator"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # Pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "ARG", # flake8-unused-arguments
    "SIM", # flake8-simplify
    "S",   # flake8-bandit (security)
    "RUF", # Ruff-specific
]
ignore = [
    "S101",  # Allow assert in tests
    "S311",  # Allow random for non-cryptographic purposes (backoff jitter, deployment names)
    "S603",  # Allow subprocess calls (required for CLI tool)
    "S607",  # Allow partial executable paths (docker, az commands)
]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["S101", "SIM117"]  # Allow assert and nested with statements in tests
"src/controller/reconciler.py" = ["SIM105"]  # Explicit try/except more readable for async waits

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[[tool.mypy.overrides]]
module = [
    "azure.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.coverage.run]
source = ["src/controller"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
]
